Correct format for labels

+++ desespoir.py
@@ -24,11 +24,11 @@ def read_hats():
             dictionary["hypB"] = line[2]
             annotation = float(line[3])
             if annotation == 0.0:
-                dictionary["annotation"] = torch.FloatTensor([1, 0])
+                dictionary["annotation"] = [1, 0]
             elif annotation == 1.0:
-                dictionary["annotation"] = torch.FloatTensor([0, 1])
+                dictionary["annotation"] = [0, 1]
             elif annotation == 0.5:
-                dictionary["annotation"] = torch.FloatTensor([0, 0]) # it is also possible to skip these cases
+                dictionary["annotation"] = [0, 0] # it is also possible to skip these cases
             else:
                 raise Exception("annotation is not 0.0, 0.5 or 1.0")
             # dictionary["annotation"] = float(line[3])
